refactor: viewModel분리
