+++ LovelyCrane/LovelyCraneApp.swift
@@ -15,7 +15,7 @@ struct LovelyCraneApp: App {
     
     var body: some Scene {
         WindowGroup {
-            RootView()
+            MainView()
         }
     }
 }
