+++ LovelyCrane/ViewModel/ConnectPage/InputCodeViewModel.swift
@@ -1,0 +1,19 @@
+//
+//  InputCodeViewModel.swift
+//  LovelyCrane
+//
+//  Created by 235 on 2023/07/25.
+//
+
+import SwiftUI
+final class InputCodeViewModel: ObservableObject {
+    @Published var inputcode: String = ""
+    func connectPartner() async throws -> Bool{
+        let failconnect = try await UserManager.shared.connectUsertoUser(to: inputcode)
+        if failconnect {
+            return false
+        }
+        inputcode = ""
+        return true
+    }
+}
